country_states_backend/country_states_backend/requirements/base.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==3.2.11
django-cors-headers==4.3.1
djangorestframework==3.14.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
psycopg2-binary==2.9.3
whitenoise==6.6.0
django-rest-knox==4.2.0